move test and user functions to scaling test dir
